+++ src/store/actions/types.ts
@@ -26,3 +26,14 @@ export interface SetTransactionsSummedAction extends ActionType {
 export interface FetchTransactionsSummedAction extends ActionType {
   transactionsSummedData: TransactionsSummedData;
 }
+
+export interface AuthUserData {
+  email: string;
+  password: string;
+  shouldSignUp: boolean
+}
+
+export interface AuthUser extends ActionType {
+  data: AuthUserData;
+}
+
